rename file and print warning
